remove duplicate api implementations
